Document how TIDs are to be returned
In https://github.com/WebAssembly/wasi-libc/pull/325, we decided that,
to avoid an extra `wait`/`notify` handshake between the parent and child
thread, the host-generated TID would be passed not only to the child
thread in a `wasi_thread_start` parameter but also returned to the
parent thread as the result of `wasi_thread_spawn`. This change
documents that decision in the spec.